refactor(services/oss): Migrate to async reqsign (#1911)
